requirements.txt
opencv-python>=4.5.0
torch>=1.9.0
torchvision>=0.10.0
ultralytics>=8.0.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0
apscheduler>=3.9.0
pillow>=8.0.0
pyserial>=3.5
requests>=2.26.0